--- app/src/main/java/com/bintina/goouttolunchmvvm/login/viewmodel/ViewModelFactory.kt
@@ -1,26 +1,0 @@
-package com.bintina.goouttolunchmvvm.login.viewmodel
-
-import android.app.Application
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.bintina.goouttolunchmvvm.login.model.database.dao.UserDao
-import com.bintina.goouttolunchmvvm.login.model.database.repositories.UserDataRepository
-import java.util.concurrent.Executor
-
-
-class ViewModelFactory(
-    val application: Application,
-    private val userDataSource: UserDataRepository,
-    private val userDao: UserDao,
-    private val executor: Executor
-) : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return if (modelClass.isAssignableFrom(UserViewModel::class.java)) {
-            UserViewModel(application, userDataSource, userDao) as T
-        } else {
-            throw IllegalArgumentException("Unknown ViewModel class")
-        }
-    }
-}
